+++ CHANGELOG.md
@@ -16,8 +16,12 @@
   - **2002 ([cleo_return_with](https://library.sannybuilder.com/#/sa/CLEO/2002))**
   - **2003 ([cleo_return_fail](https://library.sannybuilder.com/#/sa/CLEO/2003))**
   - **2004 ([forget_memory](https://library.sannybuilder.com/#/sa/CLEO/2004))**
-  - **2200 ([read_block_from_file](https://library.sannybuilder.com/#/sa/file/2200))**
+  - **2300 ([get_file_position](https://library.sannybuilder.com/#/sa/file/2300))**
+  - **2301 ([read_block_from_file](https://library.sannybuilder.com/#/sa/file/2301))**
   - opcodes **0A9A**, **0A9B**, **0A9C**, **0A9D**, **0A9E**, **0AAB**, **0AD5**, **0AD6**, **0AD7**, **0AD8**, **0AD9**, **0ADA**, **0AE4**, **0AE5**, **0AE6**, **0AE7**  and **0AE8** moved to the [FileSystemOperations](https://github.com/cleolibrary/CLEO5/tree/master/cleo_plugins/FileSystemOperations) plugin
+  - fixed bug preventing file stream opcodes from working correctly for read-write modes
+  - fixed buffer overflows in file stream read opcodes
+  - added/fixed support of all file stream opcodes in legacy mode (Cleo3)
   - 'argument count' parameter of **0AB1 (cleo_call)** is now optional. `cleo_call @LABEL args 0` can be written as `cleo_call @LABEL`
   - 'argument count' parameter of **0AB2 (cleo_return)** is now optional. `cleo_return 0` can be written as `cleo_return`
   - **cleo_return_\*** opcodes now can pass strings as return arguments